move server.js to src
