+++ ecosystem.config.js
@@ -1,6 +1,8 @@
 module.exports = {
   apps : [{
     name   : "care_why_api",
-    script : "./server.js",
+    script : "./src/server.js",
+    watch: true,
+    ignore_watch: ["upload", "database"],
   }],
 }
