--- pm2-process.json
@@ -1,9 +1,0 @@
-{
-    "script": "./server.js",
-    "watch": true,
-    "ignore_watch": ["public"],
-    "watch_options": {
-        "followSymlinks": false
-    },
-    "name": "care_why_api"
-}
